Verbose param moved to CLI package
