+++ src/cli/globalVerbose.go
@@ -1,4 +1,4 @@
-package configuration
+package cli
 
 var verboseLevel = DefaultVerbose
